+++ src/scibby/entities/Projectile.java
@@ -67,7 +67,7 @@ public abstract class Projectile extends Entity{
 	@Override
 	public void render(Screen screen){
 		if(sprite != null){	
-			screen.renderSprite((int) x, (int) y, sprite);
+			screen.renderSprite(x, y, sprite);
 		}		
 	}
